Add a type key to the returned nodes
The type key contains all labels of the host. So it would be more
correct to name it 'labels'. I chose to use type because it is
consistent with the existing naming in the library. Another option would
be to rename the other occurences of it, e.g. get_node_labels instead of
get_node_types.

Example output:

```
$ curl -s localhost:15135/neo4j/nodes/get_all  | jq .
{
  "data": [
    {
      "n": {
        "OS-DCF:diskConfig": "AUTO",
        "updated": "2019-04-13T14:57:58Z"
      },
      "type": [
        "SERVERS",
        "COMPONENT"
      ]
    }
  ]
}
```